Split the powerpoint code into service implementation and UI controller wrappers; attempting to preserve history
